# docs/swagger/swagger.yaml
definitions:
...
  routes.ErrorResponse:
    properties:
      error:
        example: not found
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/token/refresh:
    post:
      description: Accepts a valid bearer token and returns a new one with a fresh
        24h expiry.
      produces:
      - application/json
      responses:
        "200":
          description: 'New token: {\"token\":\"...\"}'
          schema:
            type: object
        "401":
          description: invalid or expired token
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh bearer token
      tags:
      - users
  /v1/user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: User with orgs array
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get authenticated user
      tags:
      - users
  /v1/users:
    post:
      consumes:
      - application/json
      description: Creates a user and optionally joins or creates an organization.
        Returns the user and a signed bearer token.
      parameters:
      - description: User creation request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: User with token field
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "409":
          description: user already exists
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
...
swagger: "2.0"